requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.7.0

# ── AI / LLM ───────────────────────────────────────────────────────────────────
groq==0.9.0

# ── Vector database + embeddings ───────────────────────────────────────────────
chromadb==0.5.3
sentence-transformers==3.0.1

# ── Cache ──────────────────────────────────────────────────────────────────────
redis==5.0.8

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-mock==3.14.0